# back/src/article/entity/article.entity.ts
import { User } from 'src/auth/entity/user.entity';
import {
  BaseEntity,
  Column,
  CreateDateColumn,
  Entity,
  ManyToOne,
  PrimaryGeneratedColumn,
} from 'typeorm';

@Entity()
export class Article extends BaseEntity {
  @PrimaryGeneratedColumn()
  id: number;

  @Column({ type: 'varchar' })
  title: string;

  @Column()
  description: string;

  @Column()
  body: string;

  @CreateDateColumn({ type: 'timestamp' })
  createdDate: Date;

  @CreateDateColumn({ type: 'timestamp' })
  updatedDate: Date;

  // TODO Скрыть это поле, не убирая из свагера тип юзера
  @ManyToOne(() => User, (user) => user.article, { eager: false })
  user: User;

  @Column({ type: 'int' })
  userId: number;
}