# back/swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/auth/signup":{"post":{"operationId":"AuthController_signUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupCredentialsDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/auth/signin":{"post":{"operationId":"AuthController_signIn","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInCredentialsDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/auth/logout":{"get":{"operationId":"AuthController_logout","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"],"security":[{"bearer":[]}]}},"/api/auth/refresh-tokens":{"post":{"operationId":"AuthController_refreshTokens","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"],"security":[{"bearer":[]}]}},"/api/article/my":{"get":{"operationId":"ArticleController_getAllArticles","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Article"}}}}}},"tags":["Article"],"security":[{"bearer":[]}]}},"/api/article":{"post":{"operationId":"ArticleController_createArticle","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArticleDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Article"}}}}},"tags":["Article"],"security":[{"bearer":[]}]}},"/api/article/{id}":{"get":{"operationId":"ArticleController_getArticleById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Article"}}}}},"tags":["Article"],"security":[{"bearer":[]}]},"patch":{"operationId":"ArticleController_updateArticleById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArticleDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Article"}}}}},"tags":["Article"],"security":[{"bearer":[]}]},"delete":{"operationId":"ArticleController_deleteArticleById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Article"],"security":[{"bearer":[]}]}},"/api/user":{"get":{"operationId":"UserController_getUser","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["User"],"security":[{"bearer":[]}]},"patch":{"operationId":"UserController_updateUserInfo","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/UserInfoDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["User"],"security":[{"bearer":[]}]}}},"info":{"title":"Articlex","description":"The Articlex API description","version":"1.0","contact":{}},"tags":[{"name":"Article","description":""}],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"SignupCredentialsDto":{"type":"object","properties":{"username":{"type":"string","minLength":4,"maxLength":20,"minimum":4,"maximum":20},"password":{"type":"string","minLength":6,"maxLength":20,"pattern":"/((?=.*\\d)|(?=.*\\w+))(?![.\\n])(?=.*[A-Z])(?=.*[a-z]).*$/","minimum":6,"maximum":20,"description":"At least 1 capital, 1 small, 1 special character and 1 number"}},"required":["username","password"]},"SignInCredentialsDto":{"type":"object","properties":{"username":{"type":"string","minLength":4,"maxLength":20,"minimum":4,"maximum":20},"password":{"type":"string","minLength":6,"maxLength":20,"pattern":"/((?=.*\\d)|(?=.*\\w+))(?![.\\n])(?=.*[A-Z])(?=.*[a-z]).*$/","minimum":6,"maximum":20,"description":"At least 1 capital, 1 small, 1 special character and 1 number"}},"required":["username","password"]},"RefreshTokenDto":{"type":"object","properties":{"refresh_token":{"type":"string"}},"required":["refresh_token"]},"Article":{"type":"object","properties":{"id":{"type":"number"},"title":{"type":"string"},"description":{"type":"string"},"body":{"type":"string"},"createdDate":{"format":"date-time","type":"string"},"updatedDate":{"format":"date-time","type":"string"},"user":{"$ref":"#/components/schemas/User"},"userId":{"type":"number"}},"required":["id","title","description","body","createdDate","updatedDate","user","userId"]},"UserInfo":{"type":"object","properties":{"id":{"type":"number"},"photo":{"type":"string"},"modified_photo":{"type":"string"}},"required":["id","photo","modified_photo"]},"User":{"type":"object","properties":{"id":{"type":"number"},"hashedRefreshToken":{"type":"string"},"salt":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"},"article":{"type":"array","items":{"$ref":"#/components/schemas/Article"}},"user_info":{"$ref":"#/components/schemas/UserInfo"}},"required":["id","salt","username","password","article","user_info"]},"ArticleDto":{"type":"object","properties":{"title":{"type":"string","minLength":4,"maxLength":30,"minimum":4,"maximum":30},"description":{"type":"string","minLength":4,"maxLength":150,"minimum":4,"maximum":150},"body":{"type":"string"}},"required":["title","description","body"]},"UserInfoDto":{"type":"object","properties":{"photo":{"type":"file","properties":{"file":{"type":"string","format":"binary"}}},"modified_photo":{"type":"string"}}}}}}